Pebble-flavoured GDB
Build a GDB which is better tailored to debugging Pebble firmware. It
links against the homebrew-dupes libiconv (fixing PBL-17248) and the
Homebrew Python install by default, and applies a patch so that GDB can
unwind past exceptions. It installs as `pebble-gdb` so as not to
conflict with other installations of GDB.